fix(test): run test running out of repository
The introduction of new typescript rules lead to the discovery of
several shortcomings in the testing process as well as
overlooked wrong test control data.

Some of the rules require `parserOptions.project`.
That option only belongs into the plugin scenario,
but was demanded by all of them.

That showed that the source repo's node_modules
was bleeding through to the tests.
In hindsight, that's just how node's module resolution works,
but went unnoticed until that point.

Consequently, the tests need to be executed in a
separate context without cross-contamination.
To achieve that, the scenarios are now setup
in a temp directory and the module code is copied over.

That in turn showed that the test control data
has been somewhat wrong for a long time.
Again in hindsight, it makes sense, but
it's unfortunate that the wrong rule ids went unnoticed.

This is fixed now, but shows the importance of proper test isolation.